Typing the 255 length to make a test case for a bug just fixed
If you allow the default to be used, 255 is an int. If you type it,
it's a string, and the Validator is responsible for converting to
an integer. This change is to type in 255, so that the test proves
this conversion is done correctly (if it's not, the annotation generates
improperly).