Fixing a bug in the tests where too many deps might be installed
This could create a situation where we tell a user to install libraries
A, B & C only, but really, we also depend on library D, which was
not recommended because we are missing that dependency from our Maker.
However, library D may have been installed in the functional test
(because we previously installed ALL dependencies, even if it was
related to a class we already had).