Add new make:dto command
This command is added as a supplement to make:form.
To prevent problems with invalid state of entities when used in forms,
data transfer objects are a solution.
This command generates these from existing entities.
It can generate getters/setters and
helper methods to fill and extract data from/to entities.

Closes #162