minor clarification about argument name
When we're looping over the service config keys, they are not really "methods",
they are service config keys. This clarifies that. Then, set determine the
method *based* from that key.

This is super minor, but it makes the code read more accurately: we are looping
over the config keys and THEN determining the php "method" based on that.