moving the reosource tests into the main test case
I realized that they didn't need to be separated - the tests did
the same thing. However, I was previously mistaken about sending only
the contents of the yaml and php to the main test method. I see now
that the original file paths are important so that we can keep the
files in the same place so that the imports work.