Refactoring how we handle service args to be very specific
There may be more optimized ways to write this algorithm, but this
tells the story more clearly: we look at the service option, then
one-by-one (via the switch) determine which option it is, and handle
that option specifically.

This is less flexible than the previous version, which may have caught
extra cases that are not in the switch-case yet. But that's a *good* thing.
I'd rather a user get a clear error that we forgot to add support for
a service option vs the class *trying* to write code for a config key
that it never expected (and likely doing it incorrectly, and not even
knowing it is incorrect to tell the user).