Introducing make:docker:database + fixing a YamlSourceManipulator bug.
Fixing bug with YamlSourceManipulator related to comments matching values

When we are walking an existing YAML structure, we try to advance to beyond
a value, but YSM got confused because a comment containing that exact string
value appeared before the value. This caused YSM to incorrectly put its pointer
in the middle of that comment, instead of advancing to the end of the real value.